+++ ClinicaFrba-BETTER_CALL_JUAN/ClinicaFrba/Modelo/Objetos/Especialidad.cs
@@ -8,8 +8,8 @@ namespace ClinicaFrba
 {
     class Especialidad
     {
-        public int codigo { get; set; }
+        public decimal codigo { get; set; }
         public string descripcion { get; set; }
-        public int tipoEspecialidadCod { get; set; }
+        public decimal tipoEspecialidadCod { get; set; }
     }
 }
